refactor(focus): extract throttle and AX-dump from FocusSnapshotResolver
Pull the deep-walk throttle (DeepGeometryWalkThrottle, now unit-tested) and the debug AX-tree dump (AXTreeDumpWriter, gated behind -cotabby-debug) into their own files; both are off the live suggestion path. FocusSnapshotResolver 1020 -> 828 lines. The live candidate-search and caret-geometry split is deferred to a stacked PR that needs manual AX smoke testing.
